requirements.txt
cbor2==5.2.0
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.10
idna==2.10
luma.core==2.2.0
luma.led-matrix==1.5.0
Pillow==8.1.0
pyftdi==0.52.0
pyserial==3.5
pyusb==1.1.0
requests==2.25.1
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
smbus2==0.4.0
spidev==3.5
urllib3==1.26.2
wrapt==1.12.1
ws2812==1.0.0
